+++ GUI/MainInstall.cs
@@ -19,9 +19,6 @@ namespace CKAN
         // this may happen on the recommended/suggested mods dialogs
         private volatile bool installCanceled;
 
-        // this will be the final list of mods we want to install
-        private HashSet<CkanModule> toInstall = new HashSet<CkanModule>();
-
         /// <summary>
         /// Initiate the GUI installer flow for one specific module
         /// </summary>
@@ -90,7 +87,8 @@ namespace CKAN
             installer.onReportModInstalled += OnModInstalled;
             // setup progress callback
 
-            toInstall       = new HashSet<CkanModule>();
+            // this will be the final list of mods we want to install
+            HashSet<CkanModule> toInstall = new HashSet<CkanModule>();
             var toUninstall = new HashSet<string>();
             var toUpgrade   = new HashSet<string>();
 
@@ -120,13 +118,13 @@ namespace CKAN
             {
                 if (change.ChangeType == GUIModChangeType.Install)
                 {
-                    AddMod(change.Mod.ToModule().recommends, recommended, change.Mod.Identifier, registry);
-                    AddMod(change.Mod.ToModule().suggests,   suggested,   change.Mod.Identifier, registry);
+                    AddMod(change.Mod.ToModule().recommends, recommended, change.Mod.Identifier, registry, toInstall);
+                    AddMod(change.Mod.ToModule().suggests,   suggested,   change.Mod.Identifier, registry, toInstall);
                 }
             }
 
-            ShowSelection(recommended);
-            ShowSelection(suggested, true);
+            ShowSelection(recommended, toInstall);
+            ShowSelection(suggested,   toInstall, true);
 
             tabController.HideTab("ChooseRecommendedModsTabPage");
 
@@ -294,80 +292,6 @@ namespace CKAN
             }
         }
 
-        private void AddMod(
-            IEnumerable<RelationshipDescriptor>  relations,
-            Dictionary<CkanModule, List<string>> chooseAble,
-            string                               identifier,
-            IRegistryQuerier                     registry)
-        {
-            if (relations == null)
-                return;
-            foreach (RelationshipDescriptor rel in relations)
-            {
-                List<CkanModule> providers = registry.LatestAvailableWithProvides(
-                    rel.name,
-                    CurrentInstance.VersionCriteria(),
-                    rel
-                );
-                foreach (CkanModule provider in providers)
-                {
-                    if (!registry.IsInstalled(provider.identifier)
-                        && !toInstall.Any(m => m.identifier == provider.identifier))
-                    {
-                        // We want to show this mod to the user. Add it.
-                        List<string> dependers;
-                        if (chooseAble.TryGetValue(provider, out dependers))
-                        {
-                            // Add the dependent mod to the list of reasons this dependency is shown.
-                            dependers.Add(identifier);
-                        }
-                        else
-                        {
-                            // Add a new entry if this provider isn't listed yet.
-                            chooseAble.Add(provider, new List<string>() { identifier });
-                        }
-                    }
-                }
-            }
-        }
-
-        private void ShowSelection(Dictionary<CkanModule, List<string>> selectable, bool suggest = false)
-        {
-            if (installCanceled)
-                return;
-
-            // If we're going to install something anyway, then don't list it in the
-            // recommended list, since they can't de-select it anyway.
-            // The ToList dance is because we need to modify the dictionary while iterating over it,
-            // and C# doesn't like that.
-            foreach (CkanModule module in selectable.Keys.ToList())
-            {
-                if (toInstall.Any(m => m.identifier == module.identifier))
-                {
-                    selectable.Remove(module);
-                }
-            }
-
-            Dictionary<CkanModule, string> mods = GetShowableMods(selectable);
-
-            // If there are any mods that would be recommended, prompt the user to make
-            // selections.
-            if (mods.Any())
-            {
-                Util.Invoke(this, () => UpdateRecommendedDialog(mods, suggest));
-
-                tabController.ShowTab("ChooseRecommendedModsTabPage", 3);
-                tabController.SetTabLock(true);
-
-                lock (this)
-                {
-                    Monitor.Wait(this);
-                }
-
-                tabController.SetTabLock(false);
-            }
-        }
-
         private void OnModInstalled(CkanModule mod)
         {
             AddStatusMessage("Module \"{0}\" successfully installed", mod.name);
@@ -485,118 +409,5 @@ namespace CKAN
             }
         }
 
-        /// <summary>
-        /// Tries to get every mod in the Dictionary, which can be installed
-        /// It also transforms the Recommender list to a string
-        /// </summary>
-        /// <param name="mods"></param>
-        /// <returns></returns>
-        private Dictionary<CkanModule, string> GetShowableMods(Dictionary<CkanModule, List<string>> mods)
-        {
-            Dictionary<CkanModule, string> modules = new Dictionary<CkanModule, string>();
-
-            var opts = new RelationshipResolverOptions
-            {
-                with_all_suggests              = false,
-                with_recommends                = false,
-                with_suggests                  = false,
-                without_enforce_consistency    = false,
-                without_toomanyprovides_kraken = true
-            };
-
-            foreach (var pair in mods)
-            {
-                try
-                {
-                    RelationshipResolver resolver = new RelationshipResolver(
-                        new List<CkanModule> { pair.Key },
-                        null,
-                        opts,
-                        RegistryManager.Instance(manager.CurrentInstance).registry,
-                        CurrentInstance.VersionCriteria()
-                    );
-
-                    if (resolver.ModList().Any())
-                    {
-                        // Resolver was able to find a way to install, so show it to the user
-                        modules.Add(pair.Key, String.Join(",", pair.Value.ToArray()));
-                    }
-                }
-                catch { }
-            }
-            return modules;
-        }
-
-        private void UpdateRecommendedDialog(Dictionary<CkanModule, string> mods, bool suggested = false)
-        {
-            if (!suggested)
-            {
-                RecommendedDialogLabel.Text =
-                    "The following modules have been recommended by one or more of the chosen modules:";
-                RecommendedModsListView.Columns[1].Text = "Recommended by:";
-                RecommendedModsToggleCheckbox.Text = "(De-)select all recommended mods.";
-                RecommendedModsToggleCheckbox.Checked=true;
-                tabController.RenameTab("ChooseRecommendedModsTabPage", "Choose recommended mods");
-            }
-            else
-            {
-                RecommendedDialogLabel.Text =
-                    "The following modules have been suggested by one or more of the chosen modules:";
-                RecommendedModsListView.Columns[1].Text = "Suggested by:";
-                RecommendedModsToggleCheckbox.Text = "(De-)select all suggested mods.";
-                RecommendedModsToggleCheckbox.Checked=false;
-                tabController.RenameTab("ChooseRecommendedModsTabPage", "Choose suggested mods");
-            }
-
-            RecommendedModsListView.Items.Clear();
-            foreach (var pair in mods)
-            {
-                CkanModule module = pair.Key;
-                ListViewItem item = new ListViewItem()
-                {
-                    Tag = module,
-                    Checked = !suggested,
-                    Text = Manager.Cache.IsMaybeCachedZip(pair.Key)
-                        ? $"{pair.Key.name} {pair.Key.version} (cached)"
-                        : $"{pair.Key.name} {pair.Key.version} ({pair.Key.download.Host ?? ""}, {CkanModule.FmtSize(pair.Key.download_size)})"
-                };
-
-                ListViewItem.ListViewSubItem recommendedBy = new ListViewItem.ListViewSubItem() { Text = pair.Value };
-
-                item.SubItems.Add(recommendedBy);
-
-                ListViewItem.ListViewSubItem description = new ListViewItem.ListViewSubItem {Text = module.@abstract};
-
-                item.SubItems.Add(description);
-
-                RecommendedModsListView.Items.Add(item);
-            }
-        }
-
-        private void RecommendedModsContinueButton_Click(object sender, EventArgs e)
-        {
-            foreach (ListViewItem item in RecommendedModsListView.Items)
-            {
-                if (item.Checked)
-                {
-                    toInstall.Add((CkanModule) item.Tag);
-                }
-            }
-
-            lock (this)
-            {
-                Monitor.Pulse(this);
-            }
-        }
-
-        private void RecommendedModsCancelButton_Click(object sender, EventArgs e)
-        {
-            installCanceled = true;
-
-            lock (this)
-            {
-                Monitor.Pulse(this);
-            }
-        }
     }
 }
